pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "okinawa-rental-finder"
version = "1.0.0"
description = "沖縄賃貸物件スクレイピング＆適正価格推定アプリ"
requires-python = ">=3.12"
dependencies = [
    "scrapy>=2.11,<3.0",
    "requests>=2.31",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "pyyaml>=6.0",
    "streamlit>=1.39",
    "pandas>=2.2",
    "numpy>=1.26",
    "scikit-learn>=1.5",
    "plotly>=5.22",
    "python-dotenv>=1.0",
    "schedule>=1.2",
    "line-notify>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
